Refactoring model namespace classes
