+++ magnet-library/magnet-processor/src/main/java/magnet/processor/scopes/Model.kt
@@ -3,7 +3,7 @@ package magnet.processor.scopes
 import com.squareup.javapoet.ClassName
 import magnet.processor.common.CommonModel
 
-interface Model {
+class Model private constructor() {
 
     class Scope(
         val type: ClassName,
