+++ magnet-library/magnet-processor/src/test/java/magnet/processor/IndexerTest.kt
@@ -3,7 +3,7 @@ package magnet.processor
 import com.google.common.truth.Truth.assertThat
 import com.squareup.javapoet.ClassName
 import magnet.processor.registry.instances.Indexer
-import magnet.processor.registry.instances.model.Inst
+import magnet.processor.registry.instances.Model.Inst
 import org.junit.Test
 
 class IndexerTest {
